+++ buildings/gui/bulk_load_changes.py
@@ -322,6 +322,9 @@ class AddBulkLoad(BulkLoadChanges):
         self.populate_edit_comboboxes()
         self.select_comboboxes_value()
 
+        self.edit_dialog.activateWindow()
+        self.edit_dialog.btn_edit_save.setDefault(True)
+
     @pyqtSlot(int)
     def creator_feature_deleted(self, qgsfId):
         """
@@ -468,6 +471,8 @@ class EditAttribute(BulkLoadChanges):
             ls_relationships = self.remove_compared_outlines()
             if len(ls_relationships['matched']) == 0 and len(ls_relationships['related']) == 0:
                 if len(self.edit_dialog.ids) > 0:
+                    # Send signal to LIQA through edit dialog
+                    self.edit_dialog.delete_outline_saved.emit(self.edit_dialog.ids, self.edit_dialog.description_del)
                     for i in self.edit_dialog.ids:
                         # check current status of building
                         sql = bulk_load_select.bulk_load_status_id_by_outline_id
@@ -790,6 +795,8 @@ class EditGeometry(BulkLoadChanges):
 
         _, capture_method_id, _, _, _, _ = self.get_comboboxes_values()
 
+        self.edit_dialog.edit_geometry_saved.emit(self.edit_dialog.geoms.keys())
+
         for key in self.edit_dialog.geoms:
             sql = 'SELECT buildings_bulk_load.bulk_load_outlines_update_shape(%s, %s);'
             self.edit_dialog.db.execute_no_commit(
@@ -857,6 +864,9 @@ class EditGeometry(BulkLoadChanges):
             self.enable_UI_functions()
             self.populate_edit_comboboxes()
             self.select_comboboxes_value()
+
+        self.edit_dialog.activateWindow()
+        self.edit_dialog.btn_edit_save.setDefault(True)
 
     def select_comboboxes_value(self):
         """