+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Unreleased
 Added
 -----
 * buildings toolbar and bulk load adding/editing functionality added to the alter relationships frame
+* Update the error status and comment in the QA layer if bulk load outline edited (edit-geometry and delete-outline only)
 
 Changed
 -------